refactor(workspaces): open-app 探测簇外移至 workspaces/open_app.rs
- commands.rs 3493→2604；open_app.rs 889 行（open_workspace_in/candidate/macOS args/icon/probe 簇）
- pub(crate) use glob 保 command_registry 零改动；cfg(windows) 属性归位修正
- 内联测试经 super::open_app::* 继续消费私有 fn
- baseline 重生成 + 新文件登记豁免
验证：cargo check 零 error；cargo test workspaces 114 passed；rustfmt 过；gate 绿
