+++ src-tauri/src/workspaces/mod.rs
@@ -3,11 +3,13 @@ mod external_changes;
 mod files;
 mod git;
 mod macos;
+mod open_app;
 mod rewind_export;
 mod settings;
 mod worktree;
 
 pub(crate) use commands::*;
+pub(crate) use open_app::*;
 pub(crate) use external_changes::DetachedExternalChangeRuntime;
 pub(crate) use rewind_export::*;
